Created the file routes and the controller for the file logic.
Most of the methods are stubs at this points but return a "Not yet implemented" error

Created some basic views that will be useful for testing like the home one and the upload

Created error views for the not implemented yet and the internal server error

Included font awsome in the bundle

Included the HTML & Form dependencie

Changed the name of the app to BerryCloud and added dependencies in layout.blade.php

Note: There is curently an issue on download where sometimes Laravel appends and inferend extension. This will be reported on an issue before merging.

Added the html form dependencies