Rikyfocil/file controller implementation (#7)
In this commit the logic of the Fille controller is fully implemented. The following were perfomed:

Overriten delete in File to delete all its versions first in case of force delete
Created a Helper class to use in views
Added functionality to the Fille controller implementing all its methods. None returns 501 anymore
Added logistic in home controller to display a table
Added constant in permission type
Added Helper class to the service provider
Changed the delete method on version
Added support in the config file to support the Helper class
Fixed the migration in the version table
Updated diagram in the pdf
Created 403 view
created simple views for displayng and demonstrating the new capabilities
Adjusted the routes to fix some issues and to give some of them a name